requirements.txt
openai==1.61.0
feedparser==6.0.11
PyYAML==6.0.2
python-dotenv==1.0.1
requests==2.32.3
openpyxl==3.1.5
python-docx==0.8.11
